don't truncate values in toDouble() (fix for jira VELOCITY-372)
git-svn-id: https://svn.apache.org/repos/asf/jakarta/velocity/tools/trunk@280564 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/velocity/tools/generic/MathTool.java
@@ -47,7 +47,7 @@ import org.apache.commons.beanutils.PropertyUtils;
  *
  * @author <a href="mailto:nathan@esha.com">Nathan Bubna</a>
  * @author Leon Messerschmidt
- * @version $Revision: 1.11 $ $Date: 2004/10/04 01:55:17 $
+ * @version $Revision$ $Date$
  */
 public class MathTool
 {
@@ -441,7 +441,7 @@ public class MathTool
         {
             return null;
         }
-        return new Double(n.intValue());
+        return new Double(n.doubleValue());
     }
